+++ _stbt/core.py
@@ -201,9 +201,6 @@ class MatchResult(object):
     :ivar Frame frame: The video frame that was searched, as given to `match`.
 
     :ivar image: The reference image that was searched for, as given to `match`.
-
-    :ivar int timestamp: DEPRECATED. Timestamp in nanoseconds. Use ``time``
-        instead.
     """
     def __init__(
             self, time, match, region, first_pass_result, frame, image,
@@ -234,13 +231,6 @@ class MatchResult(object):
     @property
     def position(self):
         return Position(self.region.x, self.region.y)
-
-    @property
-    def timestamp(self):
-        if self.time is None:
-            return None
-        else:
-            return int(self.time * 1e9)
 
 
 def load_image(filename, flags=cv2.IMREAD_COLOR):
@@ -338,9 +328,6 @@ class TextMatchResult(object):
 
     :ivar unicode text: The text that was searched for, as given to
         `match_text`.
-
-    :ivar int timestamp: DEPRECATED. Timestamp in nanoseconds. Use ``time``
-        instead.
     """
     def __init__(self, time, match, region, frame, text):
         self.time = time
@@ -362,13 +349,6 @@ class TextMatchResult(object):
                 self.region,
                 _frame_repr(self.frame),
                 self.text))
-
-    @property
-    def timestamp(self):
-        if self.time is None:
-            return None
-        else:
-            return int(self.time * 1e9)
 
 
 def new_device_under_test_from_config(
@@ -605,11 +585,11 @@ class DeviceUnderTest(object):
 
         debug("Searching for " + template.friendly_name)
 
-        for sample, _ in self.frames(timeout_secs):
+        for frame in self.frames(timeout_secs):
             result = self.match(
-                template, frame=sample, match_parameters=match_parameters,
+                template, frame=frame, match_parameters=match_parameters,
                 region=region)
-            draw_on(sample, result, label="match(%r)" %
+            draw_on(frame, result, label="match(%r)" %
                     os.path.basename(template.friendly_name))
             yield result
 
@@ -771,7 +751,7 @@ class DeviceUnderTest(object):
                 debug("timed out: %.3f > %.3f" % (timestamp, end_time))
                 return
 
-            yield frame, int(frame.time * 1e9)
+            yield frame
             first = False
 
     def get_frame(self):